Model now runs until solidification. Fixed configuration options - has been tested with H2/H2O and H2/H2O/N2 as volatiles. Plotting also seems to work and has been updated with latest ScientificColormaps library.
