+++ .gitignore
@@ -30,4 +30,5 @@ atm_chemistry.pdf
 init_spider.opts
 petsc-double/*
 sciath/*
+example_runs
