+++ init_coupler.opts
@@ -39,7 +39,7 @@ solid_stop                  = 1
 IC_INTERIOR                 = 1
 
 # Atmosphere setting: 1: Fresh start | 3: read in partial pressures
-IC_ATMOSPHERE               = 1
+IC_ATMOSPHERE               = 3
 
 # Upper boundary condition: 4: constant heat flux boundary condition
 SURFACE_BC                  = 4  
@@ -85,15 +85,15 @@ dtswitch                    = 100000
 F_atm                       = 1.0E4
 
 # Volatile species mass relative to mantle [ppm wt]
-H2O_initial_total_abundance = 315   # > 350
-CO2_initial_total_abundance = 0
-H2_initial_total_abundance  = 0
-CH4_initial_total_abundance = 0
-CO_initial_total_abundance  = 0
-N2_initial_total_abundance  = 0
-O2_initial_total_abundance  = 0
-S_initial_total_abundance   = 0 
-He_initial_total_abundance  = 0
+# H2O_initial_total_abundance = 0  
+# CO2_initial_total_abundance = 0
+# H2_initial_total_abundance  = 0
+# CH4_initial_total_abundance = 0
+# CO_initial_total_abundance  = 0
+# N2_initial_total_abundance  = 0
+# O2_initial_total_abundance  = 0
+# S_initial_total_abundance   = 0 
+# He_initial_total_abundance  = 0
 
 # Fractional species melt phase change event trigger [wt]
 H2O_poststep_change         = 0.05  
@@ -106,18 +106,18 @@ O2_poststep_change          = 0.05
 S_poststep_change           = 0.05
 He_poststep_change          = 0.05 
 
-# Optional: restart partial pressure [Pa]
-H2O_initial_atmos_pressure  = 0.0  
+# Initial atmosphere partial pressures (SPIDER) [Pa]
+H2O_initial_atmos_pressure  = 0.1e5 
 CO2_initial_atmos_pressure  = 0.0
-H2_initial_atmos_pressure   = 0.0
+H2_initial_atmos_pressure   = 0.3e5
 CH4_initial_atmos_pressure  = 0.0
 CO_initial_atmos_pressure   = 0.0
-N2_initial_atmos_pressure   = 0.0
+N2_initial_atmos_pressure   = 9.0e5
 O2_initial_atmos_pressure   = 0.0
 S_initial_atmos_pressure    = 0.0 
 He_initial_atmos_pressure   = 0.0
 
 # Atmosphere chemistry switch: 
-# 0: none 1: VULCAN->SOCRATES & 2: V->SOC+SPI
-use_vulcan                  = 2
+# 0: no chemistry, 1: use vulcan for chemistry
+use_vulcan                  = 0
