--- utils/blackbody50.opts
@@ -1,153 +1,0 @@
-# simple example SPIDER model
-# many parameter choices are available, see parameters.c for all options
-# all units are SI unless indicated otherwise
-
-#==============
-# TIME STEPPING
-#==============
-
-# number of macro steps
--nstepsmacro 12
-
-# macro step size in years
--dtmacro 100
-
-# timestepper tolerances
--ts_sundials_atol 1.0e-8
--ts_sundials_rtol 1.0e-8
-
-# number of basic nodes
--n 50
-
-# scaling parameters
-# - these are used to non-dimensionalise (scale) the numerical problem
-#   and do not represent the physical quantities that are actually
-#   used (physical parameters are listed below)
-
-# these values are chosen to roughly normalise S, dS/dt, dSdr, and d/dt(dSdr) to unity
-
--entropy0 2.60E3 # J/kg/K
--radius0 1.0E8 # m
--time0 3.154E6 #7 # s
-
-# output directory
--outputDirectory output
-
-
-#====================
-# MATERIAL PROPERTIES
-#====================
-
-# if two phase names are specified, a composite is formed. 
-# the order must be as here (melt, solid).
--phase_names melt,solid
-
--melt_TYPE 1
--melt_alpha_filename_rel_to_src lookup_data/1TPa-dK09-elec-free/thermal_exp_melt.dat
-# below shows how to use an absolute path instead
-#-melt_alpha_filename /Users/dan/Programs/spider-dev/lookup_data/1TPa-dK09-elec-free/thermal_exp_melt.dat
--melt_cp_filename_rel_to_src lookup_data/1TPa-dK09-elec-free/heat_capacity_melt.dat
--melt_dTdPs_filename_rel_to_src lookup_data/1TPa-dK09-elec-free/adiabat_temp_grad_melt.dat
--melt_rho_filename_rel_to_src lookup_data/1TPa-dK09-elec-free/density_melt.dat
--melt_temp_filename_rel_to_src lookup_data/1TPa-dK09-elec-free/temperature_melt.dat
--melt_log10visc 2.0
--melt_activation_energy 0.0
--melt_activation_volume 0.0
--melt_cond 4.0 # conductivity of melt
--melt_phase_boundary_filename_rel_to_src lookup_data/1TPa-dK09-elec-free/liquidus_A11_H13.dat
-
--solid_TYPE 1
--solid_alpha_filename_rel_to_src lookup_data/1TPa-dK09-elec-free/thermal_exp_solid.dat
--solid_cp_filename_rel_to_src lookup_data/1TPa-dK09-elec-free/heat_capacity_solid.dat
--solid_dTdPs_filename_rel_to_src lookup_data/1TPa-dK09-elec-free/adiabat_temp_grad_solid.dat
--solid_rho_filename_rel_to_src lookup_data/1TPa-dK09-elec-free/density_solid.dat
--solid_temp_filename_rel_to_src lookup_data/1TPa-dK09-elec-free/temperature_solid.dat
--solid_log10visc 21.0
--solid_activation_energy 0.0
--solid_activation_volume 0.0
--solid_visc_ref_temp -1
--solid_visc_ref_pressure -1
--solid_visc_ref_comp -1
--solid_cond 4.0 # conductivity of solid
--solid_phase_boundary_filename_rel_to_src lookup_data/1TPa-dK09-elec-free/solidus_A11_H13.dat
-
-
-#==================
-# INITIAL CONDITION
-#==================
-
-# initial condition
--IC_INTERIOR 1
--ic_adiabat_entropy 2600.0
-# next serves as an initial guess for the entropy gradient
--ic_dsdr -1.0e-5
-
-
-#====================
-# PHYSICAL PARAMETERS
-#====================
-
-# energy fluxes and sources to include
--CONDUCTION 1 # conduction
--CONVECTION 1 # convection
--MIXING 1 # mixing (latent heat transport)
--SEPARATION 1 # gravitational separation of solid and melt phase
-
-# mixing length
-# [1] is conventional, i.e. distance from the nearest boundary
-#     which is either the surface or core-mantle boundary
-# [2] is constant, i.e. 1/4 x mantle depth, which is the mean
-#     of conventional theory.  This can help to prevent a rigid
-#     lid from forming at the surface
-
--mixing_length 2
-
-# radius of the planet
--radius 6371000.0 # Earth radius
-
-# static pressure profile derived from Adams-Williamson equation of state
-# these parameters are from fitting PREM in the lower mantle (for Earth)
--adams_williamson_rhos 4078.95095544 # surface density
--adams_williamson_beta 1.1115348931000002e-07 # beta parameter
--gravity -10.0 # gravity
-
-# eddy diffusivity
-# if negative, this value is adopted (units m^2/s)
-# if positive, this value is used to scale the internally calculated eddy diffusivity
--eddy_diffusivity_thermal 1.0
--eddy_diffusivity_chemical 1.0
-
-# material properties
--grain 1.0E-3 # grain size (m)
-
-# smoothing of material properties across liquidus and solidus
-# units of melt fraction (non-dimensional)
--matprop_smooth_width 1.0E-2
-
-# viscosity
--phi_critical 0.4 # transition melt fraction (non-dimensional)
--phi_width 0.15 # transition width (non-dimensional)
-
-
-#====================
-# BOUNDARY CONDITIONS
-#====================
-
-# core-mantle boundary condition
--CORE_BC 1 # core cooling
--core_bc_value 0.0
-
-# core parameters for CORE_BC=1
--coresize 0.55 # fractional radius of core-mantle boundary
--rho_core 10738.332568062382 # core density
--cp_core 880.0 # core heat capacity
-
-# surface boundary condition
--SURFACE_BC 1 # grey-body
--surface_bc_value 0.0
-
-# emissivity of atmosphere
--emissivity0 1.0
-
-# planetary equilibrium temperature
--teqm 273.0
