+++ utils/cpl_compare_cff.py
@@ -3,6 +3,7 @@
 # Import utils- and plot-specific modules
 from utils.modules_plot import *
 from utils.modules_utils import *
+from utils.utils_spider import MyJSON
 
 import matplotlib.gridspec as gridspec
 
@@ -206,7 +207,7 @@ def plot_atmosphere( output_dir, sub_dirs ):
             time, time_idx = find_nearest(atm_data_times, output_time)
 
             # Find planet mass and radius
-            myjson_o  = su.MyJSON( data_dir+'/{}.json'.format(data_times[0]) )
+            myjson_o  = MyJSON( data_dir+'/{}.json'.format(data_times[0]) )
             xx_radius = myjson_o.get_dict_values(['data','radius_b'])*1.0E-3
             r_planet  = np.max(xx_radius*1e3) # m
             core_mass   = myjson_o.get_dict_values(['atmosphere','mass_core'])
@@ -364,12 +365,6 @@ def plot_atmosphere( output_dir, sub_dirs ):
     # ax1.set_yticklabels([])
     ax1.set_ylim(bottom=1, top=1e-5) # , top=1e-5
     
-    try:
-        sns.set_style("ticks")
-        sns.despine()
-    except:
-        print("No seaborn.")
-
     # # Legend(s)
     legend_ax0 = ax0.legend(handles=legend_ax0_handles, loc=1, ncol=1, fontsize=fs_legend, framealpha=0.99, title=r"Volatile: $z$, $p$ weighted by $\mathcal{CF}_\mathrm{F}$") # , time $t$"
     ax0.add_artist(legend_ax0)
