+++ utils/modules_plot.py
@@ -2,6 +2,7 @@
 
 # Import utils-specific modules
 from utils.modules_utils import *
+import utils.utils_spider as su
 
 from AEOLUS.SocRadConv import surf_Planck_nu
 
@@ -10,46 +11,17 @@ from natsort import natsorted # https://pypi.python.org/pypi/natsort
 # Define Crameri colormaps (+ recursive)
 from matplotlib.colors import LinearSegmentedColormap
 
+sci_colormaps = {}
 for name in [ 'acton', 'bamako', 'batlow', 'berlin', 'bilbao', 'broc', 'buda',
            'cork', 'davos', 'devon', 'grayC', 'hawaii', 'imola', 'lajolla',
            'lapaz', 'lisbon', 'nuuk', 'oleron', 'oslo', 'roma', 'tofino',
            'tokyo', 'turku', 'vik' ]:
     file = os.path.join(str(pathlib.Path(__file__).parent.absolute())+"/../AEOLUS/plotting_tools/colormaps/", name + '.txt')
     cm_data = np.loadtxt(file)
-    vars()[name] = LinearSegmentedColormap.from_list(name, cm_data)
-    vars()[name+"_r"] = LinearSegmentedColormap.from_list(name, cm_data[::-1])
-
-# Color definitions, https://chrisalbon.com/python/seaborn_color_palettes.html
-qgray       = "#768E95"
-qblue       = "#4283A9" # http://www.color-hex.com/color/4283a9
-qgreen      = "#62B4A9" # http://www.color-hex.com/color/62b4a9
-qred        = "#E6767A"
-qturq       = "#2EC0D1"
-qmagenta    = "#9A607F"
-qyellow     = "#EBB434"
-qgray_dark  = "#465559"
-qblue_dark  = "#274e65"
-qgreen_dark = "#3a6c65"
-qred_dark   = "#b85e61"
-qturq_dark  = "#2499a7"
-qmagenta_dark = "#4d303f"
-qyellow_dark  = "#a47d24"
-qgray_light  = "#acbbbf"
-qblue_light  = "#8db4cb"
-qgreen_light = "#a0d2cb"
-qred_light   = "#eb9194"
-qturq_light  = "#57ccda"
-qmagenta_light = "#c29fb2"
-qyellow_light = "#f1ca70"
-
-# color_cycle2 = [ "#1f77b4", "#ff7f0e", "#2ca02c", "#d62728", "#9467bd", "#8c564b", "#e377c", "#7f7f7f", "#bcbd22", "#17becf" ]
+    sci_colormaps[name]      = LinearSegmentedColormap.from_list(name, cm_data)
+    sci_colormaps[name+"_r"] = LinearSegmentedColormap.from_list(name, cm_data[::-1])
 
 from matplotlib import cm
-
-try:
-    import seaborn as sns
-except:
-    print("Seaborn package not installed.")
 
 # https://matplotlib.org/tutorials/colors/colormaps.html
 
@@ -69,7 +41,7 @@ vol_zorder  = {
     "NH3"            : 1,
 }
 
-vol_colors  = {
+dict_colors  = {
     "H2O"            : cm.get_cmap('PuBu', no_colors)(range(no_colors)),
     "CO2"            : cm.get_cmap("Reds", no_colors)(range(no_colors)),
     "H2"             : cm.get_cmap("Greens", no_colors)(range(no_colors)),
@@ -233,7 +205,6 @@ molar_mass      = {
 
 volatile_species = [ "H2O", "CO2", "H2", "CH4", "CO", "N2", "O2", "S", "He" ]
 
-
 # https://stackoverflow.com/questions/13490292/format-number-using-latex-notation-in-python
 def latex_float(f):
     float_str = "{0:.2g}".format(f)
@@ -298,3 +269,244 @@ def find_nearest(array, value):
     array   = np.asarray(array)
     idx     = (np.abs(array - value)).argmin()
     return array[idx], idx
+
+#===================================================================
+class MyFuncFormatter( object ):
+
+    '''the default function formatter from
+       matplotlib.ticker.FuncFormatter(func) only accepts two
+       arguments, which is not enough to scale an arcsinh function.
+       But by creating our own class here we can attach the scaling
+       to the object which can then be accessed in __call__'''
+
+    def __init__( self, arcsinh_scale ):
+        self.const = arcsinh_scale
+
+    def ascale( self, xx ):
+        '''map input to log-like values (scaled arcsinh)'''
+        yy = np.arcsinh( xx*self.const )
+        return yy
+
+    def _invascale( self, yy ):
+        '''map input from log-like values (inverse transform)'''
+        xx = np.sinh(yy) / self.const
+        return xx
+
+    def _sci_notation( self, num, decimal_digits=1, precision=None, exponent=None):
+        """
+        Returns a string representation of the scientific
+        notation of the given number formatted for use with
+        LaTeX or Mathtext, with specified number of significant
+        decimal digits and precision (number of decimal digits
+        to show). The exponent to be used can also be specified
+        explicitly.
+        """
+
+        # plotting zero is useful to emphasize that we are plotting both
+        # positive and negative values, e.g. for the heat fluxes
+        if num==0:
+            fmt = r"$0$"
+            return fmt
+
+        if not exponent:
+            exponent = abs(num)
+            exponent = np.log10( exponent )
+            exponent = np.floor( exponent )
+            exponent = int( exponent )
+
+        coeff = round(num / float(10**exponent), decimal_digits)
+        # sometimes, probably due to floating point precision? the coeff
+        # is not less than ten.  Correct for that here
+        if np.abs(coeff) >= 10.0:
+            coeff /= 10.0
+            exponent += 1
+        if not precision:
+            precision = decimal_digits
+
+        if coeff < 0.0:
+            fmt = r"$-10^{{{0}}}$".format(exponent)
+            #fmt= r"${{{0}}}$".format(exponent)
+        else:
+            fmt = r"$10^{{{0}}}$".format(exponent)
+
+        return fmt
+        #return r"${0:.{2}f}\cdot10^{{{1:d}}}$".format(coeff, exponent, precision)
+
+    def __call__( self, x, pos ):
+        y = self._invascale( x )
+        fmt = self._sci_notation( y, 0 )
+        return fmt
+
+
+#===================================================================
+class FigureData( object ):
+
+    def __init__( self, nrows, ncols, width, height, outname='fig',
+        times=None, units='kyr' ):
+        dd = {}
+        self.data_d = dd
+        if times:
+            dd['time_l'] = times
+            self.process_time_list()
+        if units:
+            dd['time_units'] = units
+            dd['time_decimal_places'] = 2 # hard-coded
+        dd['outname'] = outname
+
+        self.cmap = sci_colormaps['imola']
+            
+        self.set_properties( nrows, ncols, width, height )
+
+    def get_color( self, frac ):
+        return self.cmap(frac)
+
+    def get_legend_label( self, time ):
+        dd = self.data_d
+        units = dd['time_units']
+        dp = dd['time_decimal_places']
+        age = float(time)
+        if units == 'yr':
+            age = round( age, 0 )
+            label = '%d'
+        elif units == 'kyr':
+            age /= 1.0E3
+            label = '%0.1f'
+        elif units == 'Myr':
+            age /= 1.0E6
+            label = '%0.2f'
+        elif units == 'Byr' or units == 'Gyr':
+            age /= 1.0E9
+            label = '%0.2f'
+        #label = '%0.{}e'.format( dp )
+        #label = '%0.{}f'.format( dp )
+        label = label % age
+        return label
+
+    def process_time_list( self ):
+        dd = self.data_d
+        time_l = dd['time_l']
+        try:
+            time_l = [int(time_l)]
+        except ValueError:
+            time_l = [int(time) for time in time_l.split(',')]
+        self.time = time_l
+
+    def make_figure( self ):
+        dd = self.data_d
+        nrows = dd['nrows']
+        ncols = dd['ncols']
+        fig, ax = plt.subplots( nrows, ncols )
+        fig.subplots_adjust(wspace=0.3,hspace=0.3)
+        fig.set_size_inches( dd['width'], dd['height'] )
+        self.fig = fig
+        self.ax = ax
+
+    def savefig( self, num ):
+        dd = self.data_d
+        if dd['outname']:
+            outname = dd['outname'] + '.pdf'
+        else:
+            outname = 'fig{}.pdf'.format( num)
+        self.fig.savefig(outname, transparent=True, bbox_inches='tight',
+            pad_inches=0.05, dpi=dd['dpi'])
+
+    def set_cmap( self, cmap_obj ):
+        self.cmap = cmap_obj
+
+    def set_properties( self, nrows, ncols, width, height ):
+        dd = self.data_d
+        dd['nrows'] = nrows
+        dd['ncols'] = ncols
+        dd['width'] = width # inches
+        dd['height'] = height # inches
+        # TODO: breaks for MacOSX, since I don't think Mac comes
+        # with serif font.  But whatever it decides to switch to
+        # also looks OK and LaTeX-like.
+        font_d = {'family' : 'sans-serif',
+                  #'style': 'normal',
+                  #'weight' : 'bold'
+                  'serif': ['Arial'],
+                  'sans-serif': ['Arial'],
+                  'size'   : '10'}
+        mpl.rc('font', **font_d)
+        # Do NOT use TeX font for labels etc.
+        plt.rc('text', usetex=False)
+        dd['dpi'] = 300
+        dd['extension'] = 'png'
+        dd['fontsize_legend'] = 8
+        dd['fontsize_title'] = 10
+        dd['fontsize_xlabel'] = 10
+        dd['fontsize_ylabel'] = 10
+        self.make_figure()
+
+    def set_myaxes( self, ax, title='', xlabel='', xticks='',
+        ylabel='', yticks='', yrotation='', fmt='', xfmt='', xmin='', xmax='', ymin='', ymax='' ):
+        if title:
+            self.set_mytitle( ax, title )
+        if xlabel:
+            self.set_myxlabel( ax, xlabel )
+        if xticks:
+            self.set_myxticks( ax, xticks, xmin, xmax, xfmt )
+        if ylabel:
+            self.set_myylabel( ax, ylabel, yrotation )
+        if yticks:
+            self.set_myyticks( ax, yticks, ymin, ymax, fmt )
+
+    def set_mylegend( self, ax, handles, loc=4, ncol=1, TITLE=None, **kwargs ):
+        dd = self.data_d
+        fontsize = self.data_d['fontsize_legend']
+        # FIXME
+        if not TITLE:
+            legend = ax.legend(handles=handles, loc=loc, ncol=ncol, fontsize=fontsize, **kwargs )
+            #units = dd['time_units']
+            #title = r'Time ({0})'.format( units )
+        else:
+            title = TITLE
+            legend = ax.legend(title=title, handles=handles, loc=loc,
+                ncol=ncol, fontsize=fontsize, **kwargs)
+        plt.setp(legend.get_title(),fontsize=fontsize)
+
+    def set_mytitle( self, ax, title ):
+        dd = self.data_d
+        fontsize = dd['fontsize_title']
+        title = r'{}'.format( title )
+        ax.set_title( title, fontsize=fontsize )
+
+    def set_myxlabel( self, ax, label ):
+        dd = self.data_d
+        fontsize = dd['fontsize_xlabel']
+        label = r'{}'.format( label )
+        ax.set_xlabel( label, fontsize=fontsize )
+
+    def set_myylabel( self, ax, label, yrotation ):
+        dd = self.data_d
+        fontsize = dd['fontsize_ylabel']
+        if not yrotation:
+            yrotation = 'horizontal'
+        label = r'{}'.format( label )
+        ax.set_ylabel( label, fontsize=fontsize, rotation=yrotation )
+
+    def set_myxticks( self, ax, xticks, xmin, xmax, fmt ):
+        dd = self.data_d
+        if fmt:
+            xticks = fmt.ascale( np.array(xticks) )
+            ax.xaxis.set_major_formatter(
+                mpl.ticker.FuncFormatter(fmt))
+        ax.set_xticks( xticks)
+        # set x limits to match extent of ticks
+        if not xmax: xmax=xticks[-1]
+        if not xmin: xmin=xticks[0]
+        ax.set_xlim( xmin, xmax )
+
+    def set_myyticks( self, ax, yticks, ymin, ymax, fmt ):
+        dd = self.data_d
+        if fmt:
+            yticks = fmt.ascale( np.array(yticks) )
+            ax.yaxis.set_major_formatter(
+                mpl.ticker.FuncFormatter(fmt))
+        ax.set_yticks( yticks)
+        # set y limits to match extent of ticks
+        if not ymax: ymax=yticks[-1]
+        if not ymin: ymin=yticks[0]
+        ax.set_ylim( ymin, ymax )
+
